fix: rename folder to announcements
